projects/agents/requirements.txt
openai
agent-framework
azure-identity
python-dotenv
azure-search-documents
